Refactor CatSitter to run a compiled version of the Python script, making neccesary changes
The old way of interacting with the Python script involved having all of Python 3.9 installed and executing the script. Compiling the script using Pyinstaller is more efficient and easier for potential developers (imo, but I just hate installing things in random places). Credit where credit is due, the original method worked perfectly and was funny as hell