Refactor journal dir handling to allow for user profile expansion
This QoL change should eliminate the need for the user to manually input the profile name for their journal directory. The "?user_dir?" string will be replaced with the path to the user's profile, and the default installation path will handle the rest. The user should not have to engage with this at all.