Update pyinstaller build script & some small CatSitter changes
Add the -u Python argument to the Pyinstaller build script, this will allow the executable to provide unbuffered output which was the reason that all the output was just flying into the console with no warning. Finally fixing this fucking problem that is so simple in retrospect will mean I can actually attempt to debug on the Python side of things, thank the lord.

In the CatSitter I really just removed unused exception declerations, added a little log message for Python script startup, and added an important comment for doing things "the old fashioned way".

I might look into some form of pre-loading for the Python executable, to cut down on the (currently abysmal) startup time for the script.