Refactor webhook photos to improve error handling, streamline execution, and fix bugs
The photo list is now defined on script startup and a backup URL is defined in the event of a file access issue. Processing should be cleaner/error-free. Fixed the post_to_discord function not having a .execute() call, which led to the webhook not actually being posted. Streamlined the update_fields function, to manage the carrier and maintenance stages with much more elegance & efficiency